requirements.txt

django
djangorestframework
django-cors-headers
requests